move BpmnElementSelector to registry production code
