+++ src/component/registry/bpmn-elements-registry.ts
@@ -85,6 +85,22 @@ class BpmnModelRegistry {
   }
 }
 
+export class BpmnElementSelector {
+  constructor(private containerId: string) {}
+
+  // TODO do we make explicit that this is a svg group?
+  firstAvailableElement(bpmnElementId?: string): string {
+    if (!bpmnElementId) {
+      return `#${this.containerId} > svg > g > g > g[data-bpmn-id]`;
+    }
+    return `#${this.containerId} svg g g[data-bpmn-id="${bpmnElementId}"]`;
+  }
+
+  labelOfFirstAvailableElement(bpmnElementId?: string): string {
+    return `#${this.containerId} svg g g[data-bpmn-id="${bpmnElementId}"] g foreignObject`;
+  }
+}
+
 // TODO rename into HtmlElementRegistry
 class _HtmlElementRegistry {
   constructor(private containerId: string) {}
